requirements.txt
flask
flask-socketio
flask-jwt-extended
flask-cors
pymongo
marshmallow
python-dotenv
python-socketio[asyncio_client]
PyJWT
